requirements.txt
streamlit>=1.14.0
pandas>=1.3.0
numpy>=1.20.0
plotly>=5.3.0
pillow>=9.0.0